Moved the ipfs stuff away from the backend.js
