+++ src/services/index.js
@@ -1,1 +1,2 @@
 export * from './preview.service';
+export * from './ipfs.service';
